Revert "net: xilinx: axiethernet: Reduce the width of variable to handle overflow"
This reverts commit 3adb7e02cea0b383e54c2b6e14e5ca2068e47f14.

Non-functional coverity fixes are discouraged in mainline, so inline with
the suggestion revert coverity fix done just to make tool happy.

+++ drivers/net/ethernet/xilinx/xilinx_axienet.h
@@ -820,7 +820,7 @@ struct axienet_local {
 	u32 options;
 	u32 features;
 
-	u16 tx_bd_num;
+	u32 tx_bd_num;
 	u32 rx_bd_num;
 
 	u32 max_frm_size;
